English book reader/requirements.txt
Flask>=2.3
requests>=2.28
EbookLib>=0.18
beautifulsoup4>=4.12
lxml>=4.9
fpdf2>=2.7
genanki>=0.13
